# storage/api-docs/hrm-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mini ERP HRM API",
        "description": "Human Resources Management API Documentation for Mini ERP App - Includes Employee Management, Attendance, Leave, Payroll, Reimbursements, Resignations, Office Locations, and Analytics & Reporting",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/platform/hrm/resignations": {
            "get": {
                "tags": [
                    "HRM Resignations"
                ],
                "summary": "List resignations",
                "operationId": "a132c21b034f608c6a22d1e3bdc86915",
                "parameters": [
                    {
                        "name": "employee_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "approved",
                                "rejected",
                                "withdrawn"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "HRM Resignations"
                ],
                "summary": "Submit a resignation request",
                "operationId": "ee743b50d0b6e7d44642458f0dbd37a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "notice_date",
                                    "resignation_date",
                                    "reason"
                                ],
                                "properties": {
                                    "notice_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "resignation_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "reason": {
                                        "type": "string"
                                    },
                                    "handover_to": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resignation submitted"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/platform/hrm/resignations/{id}": {
            "get": {
                "tags": [
                    "HRM Resignations"
                ],
                "summary": "Get resignation details",
                "operationId": "9e7276da8004843ee9f67859c433c36a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/platform/hrm/resignations/{id}/status": {
            "put": {
                "tags": [
                    "HRM Resignations"
                ],
                "summary": "Update resignation status",
                "operationId": "8c2aebb0add5357b07734c689df49b8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "approved",
                                            "rejected",
                                            "withdrawn"
                                        ]
                                    },
                                    "remarks": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/platform/hrm/salary-components": {
            "get": {
                "tags": [
                    "HRM Salary Components"
                ],
                "summary": "List all salary components",
                "operationId": "e84ab5ce372d1673cc55bd6a60c4de37",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "HRM Salary Components"
                ],
                "summary": "Create a new salary component",
                "operationId": "462558402a8cdf208c0e35b45cac163a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "value"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "earning",
                                            "deduction"
                                        ]
                                    },
                                    "is_taxable": {
                                        "type": "boolean"
                                    },
                                    "is_fixed": {
                                        "type": "boolean"
                                    },
                                    "value": {
                                        "type": "number"
                                    },
                                    "percentage_of": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Component created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use a token from Sanctum auth",
                "name": "Token Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}